+++ apps/brain2print/CHANGELOG.md
@@ -1,5 +1,12 @@
 # brain2print
 
+## 0.1.20260914
+
+### Patch Changes
+
+- Updated dependencies
+  - @neurodesk/webapp-components@0.1.4
+
 ## 0.1.20260911
 
 ### Minor Changes