+++ apps/browserqc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # browserqc
 
+## 1.1.20260914
+
+### Patch Changes
+
+- Updated dependencies
+  - @neurodesk/webapp-components@0.1.4
+
 ## 1.1.20260910
 
 ### Minor Changes