+++ apps/synthseg/CHANGELOG.md
@@ -1,5 +1,13 @@
 # synthseg
 
+## 0.2.20260914
+
+### Patch Changes
+
+- Updated dependencies
+  - @neurodesk/webapp-components@0.1.4
+  - @neurodesk/synthseg@0.2.20260914
+
 ## 0.2.20260910
 
 ### Minor Changes